Add drop logic

+++ android-client/app/src/main/java/com/artemglotov/switchcase/ui/skin/viewModels/SkinViewModel.kt
@@ -1,11 +1,8 @@
 package com.artemglotov.switchcase.ui.skin.viewModels
 
 import androidx.lifecycle.*
-import com.artemglotov.switchcase.core.models.Case
 import com.artemglotov.switchcase.core.models.Skin
 import com.artemglotov.switchcase.networking.NetworkService
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 
 class SkinViewModel(
     val skin: Skin,
@@ -15,7 +12,7 @@ class SkinViewModel(
 
     }
 
-    fun getMoneyBack(): Int {
-        return skin.price.toInt()
+    fun getMoneyBack(): Float {
+        return skin.price
     }
 }
